message_broker/requirements.txt
aio-pika==9.5.0
aiormq==6.8.1
annotated-types==0.7.0
anyio==4.6.2.post1
click==8.1.7
exceptiongroup==1.2.2
fastapi==0.115.5
h11==0.14.0
idna==3.10
multidict==6.1.0
pamqp==3.3.0
pika==1.3.2
propcache==0.2.0
pydantic==2.10.2
pydantic-settings==2.6.1
pydantic_core==2.27.1
python-dotenv==1.0.1
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.32.1
yarl==1.18.0
